refactor(raylib): PrimitiveRenderer 拆分 S4c——GPU Skinned 批次与共享实例化材质管线独立（Refs #1032）
- 新建 RaylibInstancedMaterialPipeline（internal）：TryResolveInstancedModelMaterial/ApplyHostMaterialMaps/BindFrameShadow/RestoreOpaqueModelState/RequireMeshNormals/PackRgba 单一来源；PBR uniform 位置改显式 RaylibPbrUniformLocations 传参，消除按 shader.id 分派的隐式分支（行为等价：各调用点本来就按目标 shader 传对应 locs）
- 新建 RaylibGpuSkinnedBatchRenderer（internal）：GPU 蒙皮 batch 存储/submit/flush/shadow flush + skinning_instanced shader 全生命周期 + 计数器；原型蒙皮 fallback 留宿主（属原型绘制）
- 宿主 facade：公开 API、诊断属性名、计数语义、异常契约不变；ApplyFrameLighting skinning 段改为转发
- 3246 → 2696 行（拆分前 3598，累计 -25%）

验收：构建 0 错误；测试 66/67（唯一失败为预存无关项）；20 场景 Gallery 截图全量回归——全部非空渲染正常，vegetation_cutout 与基线 maxdiff=1，lighting/skybox/sky_daynight/skia_overlay/decal_projection 差异均为场景内时间驱动内容（昼夜相位/HUD/贴花绘制时机），非回归
