+++ src/Client/Ludots.Raylib.Render/Rendering/RaylibPrimitiveRenderer.cs
@@ -43,7 +43,6 @@ namespace Ludots.Raylib.Render
         private IRaylibReceiverMeshProjector? _receiverMeshProjector;
         private IVisualHeightmap? _frameVisualHeightmap;
         private Shader _shader;
-        private Shader _skinningShader;
         private Material _material;
         private Material _vfxMaterial;
         private bool _vfxMaterialLoaded;
@@ -54,28 +53,19 @@ namespace Ludots.Raylib.Render
         private int _locMetallic;
         private int _locHasRoughnessMap;
         private int _locHasMetallicMap;
-        private int _locSkinningColDiffuse;
-        private int _locSkinningTint;
-        private int _locSkinningRoughness;
-        private int _locSkinningMetallic;
-        private int _locSkinningHasRoughnessMap;
-        private int _locSkinningHasMetallicMap;
-        private int _locBoneMatrices;
+        private RaylibPbrUniformLocations _instancingPbrLocs;
         private RaylibFrameLightingLocations _instancingLightingLocs;
-        private RaylibFrameLightingLocations _skinningLightingLocs;
         private RaylibFrameLighting? _frameLighting;
         private Vector3 _frameViewPos;
         private bool _hasFrameViewPos;
         private RaylibDirectionalShadowMap? _frameShadow;
         private float _frameShadowTexelWorld = 0.04f;
-        private bool _skinningShaderReady;
         private RaylibSkyIbl? _skyIbl;
         private RaylibLitModel? _immediateLit;
         private int _locSkyZenith;
         private int _locSkyGround;
         private int _locEnvSpecular;
         private RaylibShadowSamplingLocations _instancingShadowLocs;
-        private RaylibShadowSamplingLocations _skinningShadowLocs;
         private Mesh _billboardShadowMesh;
 
         private readonly List<Batch> _cubeBatches = new List<Batch>(16);
@@ -83,11 +73,10 @@ namespace Ludots.Raylib.Render
         private readonly Dictionary<long, ModelInstanceBatch> _modelInstanceBatches = new Dictionary<long, ModelInstanceBatch>();
         private readonly Dictionary<RaylibIsmRenderBridge.Bucket, ModelInstanceBatch> _staticModelInstanceBatches = new();
         private readonly Dictionary<RaylibIsmRenderBridge.Bucket, ModelInstanceBatch> _shadowInstanceBatches = new();
-        private readonly Dictionary<GpuSkinnedInstanceBatchKey, GpuSkinnedInstanceBatch> _gpuSkinnedInstanceBatches = new();
-        private readonly List<GpuSkinnedInstanceBatch> _activeGpuSkinnedInstanceBatches = new(64);
-        private bool _gpuSkinnedBatchesPreparedForShadow;
         private readonly RaylibIsmRenderBridge _ismBridge = new RaylibIsmRenderBridge();
         private readonly RaylibGpuSkinnedModelCache _gpuSkinnedModelCache;
+        private readonly RaylibInstancedMaterialPipeline _materialPipeline;
+        private readonly RaylibGpuSkinnedBatchRenderer _gpuSkinned;
         private readonly RaylibVfxRenderer _vfxRenderer;
         private readonly RaylibDecalProjectorRenderer _decalRenderer;
         private readonly RaylibVegetationCutoutRenderer _vegetationCutout = new();
@@ -99,7 +88,6 @@ namespace Ludots.Raylib.Render
         private readonly HashSet<int> _loggedTextureDiagnostics = new HashSet<int>();
         private readonly HashSet<int> _loggedBillboardDrawDiagnostics = new HashSet<int>();
         private readonly HashSet<int> _reportedMissingModelDraws = new HashSet<int>();
-        private readonly HashSet<int> _reportedInvalidInstancedMaterials = new HashSet<int>();
         private Material _proceduralMeshMaterial;
         private bool _proceduralMeshMaterialLoaded;
         private readonly int _maxModelInstancesPerDraw;
@@ -153,12 +141,7 @@ namespace Ludots.Raylib.Render
                 ApplyFrameShadow(_shader, in _instancingShadowLocs);
             }
 
-            if (_skinningShaderReady)
-            {
-                lighting.Apply(_skinningShader, in _skinningLightingLocs);
-                lighting.ApplyViewPosition(_skinningShader, in _skinningLightingLocs, viewPos);
-                ApplyFrameShadow(_skinningShader, in _skinningShadowLocs);
-            }
+            _gpuSkinned.ApplyFrameLighting(lighting, viewPos, shadow, shadowTexelWorld);
 
             if (_immediateLit != null)
             {
@@ -193,7 +176,7 @@ namespace Ludots.Raylib.Render
 
         private void BindFrameShadow(ref Material material)
         {
-            RaylibShadowSampling.BindTexture(ref material, _frameShadow);
+            RaylibInstancedMaterialPipeline.BindFrameShadow(ref material, _frameShadow);
         }
 
         public void BindReceiverMeshProjector(IRaylibReceiverMeshProjector projector)
@@ -230,6 +213,8 @@ namespace Ludots.Raylib.Render
             _diagnosticPath = Environment.GetEnvironmentVariable("LUDOTS_RAYLIB_DIAGNOSTIC_PATH");
             _maxModelInstancesPerDraw = ResolveMaxModelInstancesPerDraw();
             _gpuSkinnedModelCache = new RaylibGpuSkinnedModelCache(vfs);
+            _materialPipeline = new RaylibInstancedMaterialPipeline(_materialHostBinder);
+            _gpuSkinned = new RaylibGpuSkinnedBatchRenderer(_gpuSkinnedModelCache, _materialPipeline, _maxModelInstancesPerDraw);
             _vfxRenderer = new RaylibVfxRenderer(vfs);
             _decalRenderer = new RaylibDecalProjectorRenderer(materials, _materialHostBinder);
         }
@@ -278,6 +263,7 @@ namespace Ludots.Raylib.Render
             LastGpuSkinnedBatches = 0;
             LastGpuSkinnedMatrixBuildMs = 0d;
             LastGpuSkinnedMeshDrawMs = 0d;
+            _gpuSkinned.ResetStats();
             LastMeshVisualCount = 0;
             LastDecalVisualCount = 0;
             LastVfxVisualCount = 0;
@@ -372,7 +358,7 @@ namespace Ludots.Raylib.Render
             if (meshes == null) throw new ArgumentNullException(nameof(meshes));
 
             var span = skinnedBatch.GetSpan();
-            PrepareGpuSkinnedInstanceBatches();
+            _gpuSkinned.Prepare();
             for (int i = 0; i < span.Length; i++)
             {
                 ref readonly var item = ref span[i];
@@ -381,7 +367,7 @@ namespace Ludots.Raylib.Render
                     continue;
                 }
 
-                if (TrySubmitGpuSkinnedInstance(item, meshes, scaleMul))
+                if (_gpuSkinned.TrySubmit(in item, meshes, scaleMul))
                 {
                     continue;
                 }
@@ -396,7 +382,11 @@ namespace Ludots.Raylib.Render
                     shadow);
             }
 
-            FlushGpuSkinnedInstanceShadowBatches(shadow);
+            _gpuSkinned.FlushShadow(shadow);
+            LastGpuSkinnedInstances = _gpuSkinned.LastInstances;
+            LastGpuSkinnedBatches = _gpuSkinned.LastBatches;
+            LastGpuSkinnedMatrixBuildMs = _gpuSkinned.LastMatrixBuildMs;
+            LastGpuSkinnedMeshDrawMs = _gpuSkinned.LastMeshDrawMs;
         }
 
         private void DrawPersistentStaticLanes(Camera3D camera, IRenderMeshAssets meshes, float scaleMul)
@@ -632,7 +622,7 @@ namespace Ludots.Raylib.Render
 
         private void SubmitPrimitive(PrimitiveMeshKind kind, Vector3 position, Quaternion rotation, Vector3 scale, Vector4 color, int materialId = 0)
         {
-            uint key = PackRgba(color);
+            uint key = RaylibInstancedMaterialPipeline.PackRgba(color);
             var matrix = RaylibMatrix.FromSystemNumerics(
                 Matrix4x4.CreateScale(scale) *
                 Matrix4x4.CreateFromQuaternion(VisualMath.NormalizeOrIdentity(rotation)) *
@@ -686,10 +676,9 @@ namespace Ludots.Raylib.Render
         private void DrawSkinnedBatch(ISkinnedVisualBatchSnapshot skinnedBatch, Camera3D camera, IRenderMeshAssets meshes, float scaleMul)
         {
             var span = skinnedBatch.GetSpan();
-            bool reuseShadowBatches = _gpuSkinnedBatchesPreparedForShadow;
-            if (!reuseShadowBatches)
+            if (!_gpuSkinned.BatchesPreparedForShadow)
             {
-                PrepareGpuSkinnedInstanceBatches();
+                _gpuSkinned.Prepare();
             }
             for (int i = 0; i < span.Length; i++)
             {
@@ -699,7 +688,7 @@ namespace Ludots.Raylib.Render
                     continue;
                 }
 
-                if (TrySubmitGpuSkinnedInstance(item, meshes, scaleMul))
+                if (_gpuSkinned.TrySubmit(in item, meshes, scaleMul))
                 {
                     continue;
                 }
@@ -720,165 +709,16 @@ namespace Ludots.Raylib.Render
                     item.MaterialId);
             }
 
-            FlushGpuSkinnedInstanceBatches();
-            _gpuSkinnedBatchesPreparedForShadow = false;
+            if (_gpuSkinned.HasActiveBatches)
+            {
+                EnsureInitialized();
+            }
+            _gpuSkinned.Flush(_shader, in _instancingPbrLocs, _skyIbl);
+            LastGpuSkinnedInstances = _gpuSkinned.LastInstances;
+            LastGpuSkinnedBatches = _gpuSkinned.LastBatches;
+            LastGpuSkinnedMatrixBuildMs = _gpuSkinned.LastMatrixBuildMs;
+            LastGpuSkinnedMeshDrawMs = _gpuSkinned.LastMeshDrawMs;
         }
-
-        private void PrepareGpuSkinnedInstanceBatches()
-        {
-            for (int i = 0; i < _activeGpuSkinnedInstanceBatches.Count; i++)
-            {
-                _activeGpuSkinnedInstanceBatches[i].Count = 0;
-                _activeGpuSkinnedInstanceBatches[i].BonesPrepared = false;
-            }
-
-            _activeGpuSkinnedInstanceBatches.Clear();
-        }
-
-        private bool TrySubmitGpuSkinnedInstance(in SkinnedVisualBatchItem item, IRenderMeshAssets meshes, float scaleMul)
-        {
-            if (item.RenderPath != VisualRenderPath.GpuSkinnedInstance ||
-                !meshes.TryGetDescriptor(item.MeshAssetId, out MeshAssetDescriptor descriptor) ||
-                descriptor.Type != MeshAssetType.Model)
-            {
-                return false;
-            }
-
-            RaylibGpuSkinnedModelCache.Entry entry = _gpuSkinnedModelCache.GetOrLoad(item.MeshAssetId, in descriptor);
-            AnimatorPackedState animator = item.Animator;
-            RaylibSkinnedPlayback.ResolveFromAnimator(
-                in animator,
-                entry.Animations,
-                entry.AnimCount,
-                stateToClipMap: null,
-                out int clipIndex,
-                out int frameIndex);
-
-            long start = Stopwatch.GetTimestamp();
-            uint colorKey = PackRgba(item.Color);
-            var key = new GpuSkinnedInstanceBatchKey(
-                item.MeshAssetId,
-                item.MaterialId,
-                colorKey,
-                clipIndex,
-                frameIndex);
-            if (!_gpuSkinnedInstanceBatches.TryGetValue(key, out GpuSkinnedInstanceBatch? batch))
-            {
-                batch = new GpuSkinnedInstanceBatch(key);
-                _gpuSkinnedInstanceBatches.Add(key, batch);
-            }
-
-            if (batch.Count == 0)
-            {
-                batch.Model = entry.Model;
-                batch.Animations = entry.Animations;
-                batch.AnimCount = entry.AnimCount;
-                _activeGpuSkinnedInstanceBatches.Add(batch);
-            }
-
-            if (_gpuSkinnedBatchesPreparedForShadow && batch.Count > 0)
-            {
-                return true;
-            }
-
-            batch.Add(RaylibMatrix.FromSystemNumerics(
-                Matrix4x4.CreateScale(item.Scale * scaleMul) *
-                Matrix4x4.CreateFromQuaternion(VisualMath.NormalizeOrIdentity(item.Rotation)) *
-                Matrix4x4.CreateTranslation(item.Position)));
-            LastGpuSkinnedMatrixBuildMs += (Stopwatch.GetTimestamp() - start) * 1000d / Stopwatch.Frequency;
-            return true;
-        }
-
-        private void FlushGpuSkinnedInstanceBatches()
-        {
-            if (_activeGpuSkinnedInstanceBatches.Count == 0)
-            {
-                return;
-            }
-
-            EnsureInitialized();
-            EnsureSkinningShaderInitialized();
-            long drawStart = Stopwatch.GetTimestamp();
-            for (int i = 0; i < _activeGpuSkinnedInstanceBatches.Count; i++)
-            {
-                GpuSkinnedInstanceBatch batch = _activeGpuSkinnedInstanceBatches[i];
-                if (batch.Count == 0)
-                {
-                    continue;
-                }
-
-                LastGpuSkinnedInstances += batch.Count;
-                LastGpuSkinnedBatches += DrawGpuSkinnedInstanceBatch(batch);
-            }
-
-            LastGpuSkinnedMeshDrawMs += (Stopwatch.GetTimestamp() - drawStart) * 1000d / Stopwatch.Frequency;
-        }
-
-        private void FlushGpuSkinnedInstanceShadowBatches(RaylibDirectionalShadowMap shadow)
-        {
-            if (_activeGpuSkinnedInstanceBatches.Count == 0)
-            {
-                return;
-            }
-
-            for (int i = 0; i < _activeGpuSkinnedInstanceBatches.Count; i++)
-            {
-                GpuSkinnedInstanceBatch batch = _activeGpuSkinnedInstanceBatches[i];
-                if (batch.Count == 0)
-                {
-                    continue;
-                }
-
-                DrawGpuSkinnedInstanceBatchShadow(batch, shadow);
-            }
-
-            _gpuSkinnedBatchesPreparedForShadow = true;
-        }
-
-        private void DrawGpuSkinnedInstanceBatchShadow(GpuSkinnedInstanceBatch batch, RaylibDirectionalShadowMap shadow)
-        {
-            Model model = batch.Model;
-            if (model.meshCount <= 0 || batch.Count <= 0)
-            {
-                return;
-            }
-
-            if (batch.Animations == null || batch.AnimCount <= 0)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} GpuSkinned shadow batch meshAssetId={batch.Key.MeshAssetId} has no animations; silent static shadow is forbidden.");
-            }
-
-            int clipIndex = batch.Key.ClipIndex;
-            int frameIndex = batch.Key.FrameIndex;
-            if ((uint)clipIndex >= (uint)batch.AnimCount)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} GpuSkinned shadow batch clipIndex={clipIndex} outside animCount={batch.AnimCount}.");
-            }
-
-            ModelAnimation anim = batch.Animations[clipIndex];
-            Rl.UpdateModelAnimationBones(model, anim, frameIndex);
-            batch.BonesPrepared = true;
-            fixed (RaylibMatrix* transforms = batch.Transforms)
-            {
-                for (int meshIndex = 0; meshIndex < model.meshCount; meshIndex++)
-                {
-                    Mesh mesh = model.meshes[meshIndex];
-                    if (mesh.vertexCount <= 0)
-                    {
-                        continue;
-                    }
-
-                    for (int offset = 0; offset < batch.Count; offset += _maxModelInstancesPerDraw)
-                    {
-                        int chunkCount = Math.Min(_maxModelInstancesPerDraw, batch.Count - offset);
-                        shadow.DrawSkinnedMeshInstancedShadow(mesh, transforms + offset, chunkCount);
-                    }
-                }
-            }
-        }
-
         private bool TryDrawPrototypeSkinned(in SkinnedVisualBatchItem item, IRenderMeshAssets meshes, float scaleMul)
         {
             if (!item.RenderPath.IsSkinnedLane() ||
@@ -1382,7 +1222,7 @@ namespace Ludots.Raylib.Render
             var model = cached.Model;
             ApplyHostMapsToModel(ref model, materialId);
             ToAxisAngleDegrees(rotation, out Vector3 axis, out float angleDegrees);
-            RestoreOpaqueModelState();
+            RaylibInstancedMaterialPipeline.RestoreOpaqueModelState();
             Rl.DrawModelEx(model, position, axis, angleDegrees, scale, tint);
         }
 
@@ -1575,7 +1415,7 @@ namespace Ludots.Raylib.Render
         private Material ResolveProceduralDrawMaterial(int materialAssetId)
         {
             Material material = _proceduralMeshMaterial;
-            ApplyHostMaterialMaps(ref material, materialAssetId, _proceduralMeshMaterial.shader);
+            _materialPipeline.ApplyHostMaterialMaps(ref material, materialAssetId, _proceduralMeshMaterial.shader, in _instancingPbrLocs);
             BindFrameShadow(ref material);
             return material;
         }
@@ -2202,7 +2042,7 @@ namespace Ludots.Raylib.Render
                 return true;
             }
 
-            uint colorKey = PackRgba(first.Color);
+            uint colorKey = RaylibInstancedMaterialPipeline.PackRgba(first.Color);
             long batchKey = BuildModelInstanceBatchKey(first.MeshAssetId, colorKey);
             bool canCacheStaticMatrices = bucket.Lane.Mobility == VisualMobility.Static && MathF.Abs(scaleMul - 1f) <= 0.0001f;
             ModelInstanceBatch batch;
@@ -2371,14 +2211,14 @@ namespace Ludots.Raylib.Render
             EnsureFrameLightingAppliedForInstancing();
             int drawCalls = 0;
             long drawStart = Stopwatch.GetTimestamp();
-            RestoreOpaqueModelState();
+            RaylibInstancedMaterialPipeline.RestoreOpaqueModelState();
             fixed (RaylibMatrix* transforms = batch.Transforms)
             {
                 for (int meshIndex = 0; meshIndex < model.meshCount; meshIndex++)
                 {
                     Mesh mesh = model.meshes[meshIndex];
-                    RequireMeshNormals(in mesh, "Instanced ISM");
-                    if (!TryResolveInstancedModelMaterial(model, meshIndex, materialId, out Material material))
+                    RaylibInstancedMaterialPipeline.RequireMeshNormals(in mesh, "Instanced ISM");
+                    if (!_materialPipeline.TryResolveInstancedModelMaterial(model, meshIndex, materialId, _shader, in _instancingPbrLocs, _skyIbl, _frameShadow, out Material material))
                     {
                         continue;
                     }
@@ -2395,151 +2235,6 @@ namespace Ludots.Raylib.Render
             LastInstancedMeshDrawMs += (Stopwatch.GetTimestamp() - drawStart) * 1000.0 / Stopwatch.Frequency;
             return drawCalls;
         }
-
-        private int DrawGpuSkinnedInstanceBatch(GpuSkinnedInstanceBatch batch)
-        {
-            Model model = batch.Model;
-            if (model.meshCount <= 0 || batch.Count <= 0)
-            {
-                return 0;
-            }
-
-            if (batch.Animations == null || batch.AnimCount <= 0)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} GpuSkinned batch meshAssetId={batch.Key.MeshAssetId} has no animations; silent static draw is forbidden.");
-            }
-
-            int clipIndex = batch.Key.ClipIndex;
-            int frameIndex = batch.Key.FrameIndex;
-            if ((uint)clipIndex >= (uint)batch.AnimCount)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} GpuSkinned batch clipIndex={clipIndex} outside animCount={batch.AnimCount}.");
-            }
-
-            ModelAnimation anim = batch.Animations[clipIndex];
-            if (!batch.BonesPrepared)
-            {
-                Rl.UpdateModelAnimationBones(model, anim, frameIndex);
-                batch.BonesPrepared = true;
-            }
-
-            EnsureFrameLightingAppliedForSkinning();
-            int drawCalls = 0;
-            uint colorKey = batch.Key.ColorKey;
-            int materialId = batch.Key.MaterialId;
-            RestoreOpaqueModelState();
-            fixed (RaylibMatrix* transforms = batch.Transforms)
-            {
-                for (int meshIndex = 0; meshIndex < model.meshCount; meshIndex++)
-                {
-                    Mesh mesh = model.meshes[meshIndex];
-                    if (mesh.vertexCount <= 0)
-                    {
-                        continue;
-                    }
-
-                    RequireMeshNormals(in mesh, "GpuSkinnedInstance");
-                    if (!TryResolveInstancedModelMaterial(model, meshIndex, materialId, out Material material))
-                    {
-                        continue;
-                    }
-
-                    material.shader = _skinningShader;
-                    ApplyHostMaterialMaps(ref material, materialId, _skinningShader);
-                    BindFrameShadow(ref material);
-                    ApplyGpuSkinnedMaterialTint(ref material, colorKey);
-
-                    if (mesh.boneMatrices != null && mesh.boneCount > 0)
-                    {
-                        Rl.rlEnableShader(_skinningShader.id);
-                        Rl.rlSetUniformMatrices(_locBoneMatrices, mesh.boneMatrices, mesh.boneCount);
-                    }
-
-                    for (int offset = 0; offset < batch.Count; offset += _maxModelInstancesPerDraw)
-                    {
-                        int chunkCount = Math.Min(_maxModelInstancesPerDraw, batch.Count - offset);
-                        Rl.DrawMeshInstanced(mesh, material, transforms + offset, chunkCount);
-                        drawCalls++;
-                    }
-                }
-            }
-
-            return drawCalls;
-        }
-
-        private static void RestoreOpaqueModelState()
-        {
-            Rl.rlEnableDepthTest();
-            Rl.rlEnableDepthMask();
-            Rl.rlEnableBackfaceCulling();
-        }
-
-        private bool TryResolveInstancedModelMaterial(Model model, int meshIndex, int materialId, out Material material)
-        {
-            if (model.materialCount <= 0 || model.materials == null)
-            {
-                material = default;
-                int reportKey = HashCode.Combine(model.meshCount, meshIndex, model.materialCount);
-                if (_reportedInvalidInstancedMaterials.Add(reportKey))
-                {
-                    RenderDiagnostics.Warn($"Raylib instanced model skipped meshIndex={meshIndex}: imported model has no material. Host asset material import must provide an explicit material.");
-                }
-
-                return false;
-            }
-
-            int materialIndex = 0;
-            if (model.meshMaterial != null && meshIndex >= 0 && meshIndex < model.meshCount)
-            {
-                materialIndex = model.meshMaterial[meshIndex];
-            }
-
-            if (materialIndex < 0 || materialIndex >= model.materialCount)
-            {
-                int reportKey = HashCode.Combine(model.meshCount, meshIndex, materialIndex);
-                if (_reportedInvalidInstancedMaterials.Add(reportKey))
-                {
-                    RenderDiagnostics.Warn($"Raylib instanced model skipped meshIndex={meshIndex}: meshMaterial index {materialIndex} is outside materialCount={model.materialCount}.");
-                }
-
-                material = default;
-                return false;
-            }
-
-            material = model.materials[materialIndex];
-            material.shader = _shader;
-            if (_skyIbl != null)
-            {
-                // maps 为材质共享指针：写入即持久挂载；重烘后 id 变化由每帧重挂覆盖。
-                Rl.SetMaterialTexture(ref material, (int)Rl.MaterialMapIndex.MATERIAL_MAP_CUBEMAP, _skyIbl.EnvCubemap);
-                Rl.SetMaterialTexture(ref material, (int)Rl.MaterialMapIndex.MATERIAL_MAP_BRDF, _skyIbl.BrdfLut);
-            }
-
-            ApplyHostMaterialMaps(ref material, materialId, _shader);
-            BindFrameShadow(ref material);
-            return true;
-        }
-
-        private void ApplyHostMaterialMaps(ref Material material, int materialId, Shader shader)
-        {
-            if (_materialHostBinder == null || materialId <= 0)
-            {
-                _materialHostBinder?.DetachOwnedMaps(ref material);
-                ApplyPbrUniforms(shader, materialId: 0, hostBound: false);
-                return;
-            }
-
-            bool hostBound = _materialHostBinder.TryApplyHostMaps(ref material, materialId);
-            if (!hostBound)
-            {
-                _materialHostBinder.DetachOwnedMaps(ref material);
-            }
-
-            ApplyPbrUniforms(shader, materialId, hostBound);
-        }
-
         private void ApplyHostMapsToModel(ref Model model, int materialId)
         {
             if (_materialHostBinder == null || materialId <= 0 || model.materialCount <= 0 || model.materials == null)
@@ -2550,58 +2245,9 @@ namespace Ludots.Raylib.Render
             for (int i = 0; i < model.materialCount; i++)
             {
                 ref Material material = ref model.materials[i];
-                ApplyHostMaterialMaps(ref material, materialId, material.shader.id != 0 ? material.shader : _shader);
+                _materialPipeline.ApplyHostMaterialMaps(ref material, materialId, material.shader.id != 0 ? material.shader : _shader, in _instancingPbrLocs);
             }
         }
-
-        private void ApplyPbrUniforms(Shader shader, int materialId, bool hostBound)
-        {
-            float roughness = RaylibMaterialHostBinder.DefaultRoughness;
-            float metallic = RaylibMaterialHostBinder.DefaultMetallic;
-            int hasRoughnessMap = 0;
-            int hasMetallicMap = 0;
-
-            if (hostBound &&
-                _materialHostBinder != null &&
-                _materialHostBinder.TryGetHostPbrParams(
-                    materialId,
-                    out roughness,
-                    out metallic,
-                    out bool hasRoughness,
-                    out bool hasMetallic,
-                    out _))
-            {
-                hasRoughnessMap = hasRoughness ? 1 : 0;
-                hasMetallicMap = hasMetallic ? 1 : 0;
-            }
-
-            bool isSkinning = _skinningShaderReady && shader.id == _skinningShader.id;
-            int locRoughness = isSkinning ? _locSkinningRoughness : _locRoughness;
-            int locMetallic = isSkinning ? _locSkinningMetallic : _locMetallic;
-            int locHasRoughness = isSkinning ? _locSkinningHasRoughnessMap : _locHasRoughnessMap;
-            int locHasMetallic = isSkinning ? _locSkinningHasMetallicMap : _locHasMetallicMap;
-
-            if (locRoughness >= 0)
-            {
-                Rl.SetShaderValue(shader, locRoughness, &roughness, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_FLOAT);
-            }
-
-            if (locMetallic >= 0)
-            {
-                Rl.SetShaderValue(shader, locMetallic, &metallic, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_FLOAT);
-            }
-
-            if (locHasRoughness >= 0)
-            {
-                Rl.SetShaderValue(shader, locHasRoughness, &hasRoughnessMap, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_INT);
-            }
-
-            if (locHasMetallic >= 0)
-            {
-                Rl.SetShaderValue(shader, locHasMetallic, &hasMetallicMap, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_INT);
-            }
-        }
-
         private void ApplyInstancedMaterialTint(ref Material material, uint colorKey)
         {
             SetTintUniform(colorKey);
@@ -2613,28 +2259,6 @@ namespace Ludots.Raylib.Render
                 Rl.SetShaderValue(_shader, _locColDiffuse, &diffuse, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_VEC4);
             }
         }
-
-        private void ApplyGpuSkinnedMaterialTint(ref Material material, uint colorKey)
-        {
-            if (_locSkinningTint >= 0)
-            {
-                float r = (colorKey & 0xFF) / 255f;
-                float g = ((colorKey >> 8) & 0xFF) / 255f;
-                float b = ((colorKey >> 16) & 0xFF) / 255f;
-                float a = ((colorKey >> 24) & 0xFF) / 255f;
-                var tint = new Vector4(r, g, b, a);
-                Rl.SetShaderValue(_skinningShader, _locSkinningTint, &tint, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_VEC4);
-            }
-
-            if (_locSkinningColDiffuse >= 0 && material.maps != null)
-            {
-                int albedoIndex = (int)Rl.MaterialMapIndex.MATERIAL_MAP_ALBEDO;
-                Color color = material.maps[albedoIndex].color;
-                Vector4 diffuse = new(color.r / 255f, color.g / 255f, color.b / 255f, color.a / 255f);
-                Rl.SetShaderValue(_skinningShader, _locSkinningColDiffuse, &diffuse, (int)Rl.ShaderUniformDataType.SHADER_UNIFORM_VEC4);
-            }
-        }
-
         private static int ResolveMaxModelInstancesPerDraw()
         {
             string? raw = Environment.GetEnvironmentVariable("LUDOTS_RAYLIB_MAX_MODEL_INSTANCES_PER_DRAW");
@@ -2678,7 +2302,7 @@ namespace Ludots.Raylib.Render
         private void FlushMeshBatches(List<Batch> batches, ref int totalInstances, ref int batchCount, ref Mesh mesh)
         {
             EnsureFrameLightingAppliedForInstancing();
-            RequireMeshNormals(in mesh, "Instanced primitive");
+            RaylibInstancedMaterialPipeline.RequireMeshNormals(in mesh, "Instanced primitive");
             for (int i = 0; i < batches.Count; i++)
             {
                 var b = batches[i];
@@ -2686,7 +2310,7 @@ namespace Ludots.Raylib.Render
 
                 SetTintUniform(b.ColorKey);
                 SetColDiffuseUniform(Vector4.One);
-                ApplyHostMaterialMaps(ref _material, b.MaterialId, _shader);
+                _materialPipeline.ApplyHostMaterialMaps(ref _material, b.MaterialId, _shader, in _instancingPbrLocs);
                 BindFrameShadow(ref _material);
 
                 fixed (RaylibMatrix* p = b.Transforms)
@@ -2770,6 +2394,7 @@ namespace Ludots.Raylib.Render
             _locMetallic = Rl.GetShaderLocation(_shader, "uMetallic");
             _locHasRoughnessMap = Rl.GetShaderLocation(_shader, "uHasRoughnessMap");
             _locHasMetallicMap = Rl.GetShaderLocation(_shader, "uHasMetallicMap");
+            _instancingPbrLocs = new RaylibPbrUniformLocations(_locRoughness, _locMetallic, _locHasRoughnessMap, _locHasMetallicMap);
             _locSkyZenith = RaylibShaderBindingGuard.RequireUniform(_shader, "uSkyZenith", "instancing");
             _locSkyGround = RaylibShaderBindingGuard.RequireUniform(_shader, "uSkyGround", "instancing");
             _locEnvSpecular = RaylibShaderBindingGuard.RequireUniform(_shader, "uEnvSpecular", "instancing");
@@ -2833,7 +2458,7 @@ namespace Ludots.Raylib.Render
             if (locMapMetalness < 0) throw new InvalidOperationException("Shader uniform 'texture1' not found.");
             if (locMapRoughness < 0) throw new InvalidOperationException("Shader uniform 'texture3' not found.");
 
-            ApplyPbrUniforms(_shader, materialId: 0, hostBound: false);
+            _materialPipeline.ApplyDefaultPbrUniforms(_shader, in _instancingPbrLocs);
 
             _initialized = true;
             ApplyFrameShadow(_shader, in _instancingShadowLocs);
@@ -2869,96 +2494,6 @@ namespace Ludots.Raylib.Render
             Rl.UploadMesh(ref mesh, false);
             return mesh;
         }
-
-        private void EnsureSkinningShaderInitialized()
-        {
-            if (_skinningShaderReady)
-            {
-                return;
-            }
-
-            string baseDir = AppContext.BaseDirectory;
-            string vsPath = Path.Combine(baseDir, "skinning_instanced.vs");
-            string fsPath = Path.Combine(baseDir, "skinning_instanced.fs");
-            if (!File.Exists(vsPath) || !File.Exists(fsPath))
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} GpuSkinnedInstance requires skinning_instanced.vs/.fs beside the binary (missing under '{baseDir}').");
-            }
-
-            _skinningShader = RaylibShaderLoader.Load(baseDir, "skinning_instanced.vs", "skinning_instanced.fs", "skinning_instanced");
-
-            _locBoneMatrices = Rl.GetShaderLocation(_skinningShader, "boneMatrices");
-            _locSkinningTint = Rl.GetShaderLocation(_skinningShader, "tint");
-            _locSkinningColDiffuse = Rl.GetShaderLocation(_skinningShader, "colDiffuse");
-            _locSkinningRoughness = Rl.GetShaderLocation(_skinningShader, "uRoughness");
-            _locSkinningMetallic = Rl.GetShaderLocation(_skinningShader, "uMetallic");
-            _locSkinningHasRoughnessMap = Rl.GetShaderLocation(_skinningShader, "uHasRoughnessMap");
-            _locSkinningHasMetallicMap = Rl.GetShaderLocation(_skinningShader, "uHasMetallicMap");
-            int locMapAlbedo = Rl.GetShaderLocation(_skinningShader, "texture0");
-            int locMapMetalness = Rl.GetShaderLocation(_skinningShader, "texture1");
-            int locMapRoughness = Rl.GetShaderLocation(_skinningShader, "texture3");
-            int locMvp = Rl.GetShaderLocation(_skinningShader, "mvp");
-            int locInstance = Rl.GetShaderLocationAttrib(_skinningShader, "instanceTransform");
-            int locVertexPosition = Rl.GetShaderLocationAttrib(_skinningShader, "vertexPosition");
-            int locVertexTexCoord = Rl.GetShaderLocationAttrib(_skinningShader, "vertexTexCoord");
-            int locVertexNormal = Rl.GetShaderLocationAttrib(_skinningShader, "vertexNormal");
-            int locVertexColor = Rl.GetShaderLocationAttrib(_skinningShader, "vertexColor");
-            int locBoneIds = Rl.GetShaderLocationAttrib(_skinningShader, "vertexBoneIds");
-            int locBoneWeights = Rl.GetShaderLocationAttrib(_skinningShader, "vertexBoneWeights");
-            _skinningLightingLocs = RaylibFrameLightingLocations.ResolveOrThrow(_skinningShader, "skinning_instanced");
-            _skinningShadowLocs = RaylibShadowSamplingLocations.ResolveOrThrow(
-                _skinningShader,
-                "skinning_instanced",
-                RaylibShadowSampling.ShaderTextureSlot);
-
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_POSITION] = locVertexPosition;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_TEXCOORD01] = locVertexTexCoord;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_NORMAL] = locVertexNormal;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_COLOR] = locVertexColor;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_BONEIDS] = locBoneIds;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_VERTEX_BONEWEIGHTS] = locBoneWeights;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MATRIX_MVP] = locMvp;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MATRIX_MODEL] = locInstance;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_COLOR_DIFFUSE] = _locSkinningColDiffuse;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MAP_ALBEDO] = locMapAlbedo;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MAP_METALNESS] = locMapMetalness;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MAP_NORMAL] = -1;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MAP_ROUGHNESS] = locMapRoughness;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_MAP_EMISSION] = _skinningShadowLocs.ShadowMap;
-            _skinningShader.locs[(int)Rl.ShaderLocationIndex.SHADER_LOC_BONE_MATRICES] = _locBoneMatrices;
-
-            if (_locBoneMatrices < 0) throw new InvalidOperationException("Skinning shader uniform 'boneMatrices' not found.");
-            if (locMvp < 0) throw new InvalidOperationException("Skinning shader uniform 'mvp' not found.");
-            if (locInstance < 0) throw new InvalidOperationException("Skinning shader attrib 'instanceTransform' not found.");
-            if (locVertexPosition < 0) throw new InvalidOperationException("Skinning shader attrib 'vertexPosition' not found.");
-            if (locVertexNormal < 0) throw new InvalidOperationException("Skinning shader attrib 'vertexNormal' not found.");
-            if (locBoneIds < 0) throw new InvalidOperationException("Skinning shader attrib 'vertexBoneIds' not found.");
-            if (locBoneWeights < 0) throw new InvalidOperationException("Skinning shader attrib 'vertexBoneWeights' not found.");
-            if (_locSkinningColDiffuse < 0) throw new InvalidOperationException("Skinning shader uniform 'colDiffuse' not found.");
-            if (_locSkinningTint < 0) throw new InvalidOperationException("Skinning shader uniform 'tint' not found.");
-            if (locMapAlbedo < 0) throw new InvalidOperationException("Skinning shader uniform 'texture0' not found.");
-            if (_locSkinningRoughness < 0) throw new InvalidOperationException("Skinning shader uniform 'uRoughness' not found.");
-            if (_locSkinningMetallic < 0) throw new InvalidOperationException("Skinning shader uniform 'uMetallic' not found.");
-            if (_locSkinningHasRoughnessMap < 0) throw new InvalidOperationException("Skinning shader uniform 'uHasRoughnessMap' not found.");
-            if (_locSkinningHasMetallicMap < 0) throw new InvalidOperationException("Skinning shader uniform 'uHasMetallicMap' not found.");
-            if (locMapMetalness < 0) throw new InvalidOperationException("Skinning shader uniform 'texture1' not found.");
-            if (locMapRoughness < 0) throw new InvalidOperationException("Skinning shader uniform 'texture3' not found.");
-
-            ApplyPbrUniforms(_skinningShader, materialId: 0, hostBound: false);
-
-            _skinningShaderReady = true;
-            ApplyFrameShadow(_skinningShader, in _skinningShadowLocs);
-            if (_frameLighting != null)
-            {
-                _frameLighting.Apply(_skinningShader, in _skinningLightingLocs);
-                if (_hasFrameViewPos)
-                {
-                    _frameLighting.ApplyViewPosition(_skinningShader, in _skinningLightingLocs, _frameViewPos);
-                }
-            }
-        }
-
         private void EnsureFrameLightingAppliedForInstancing()
         {
             if (_frameLighting == null)
@@ -2978,27 +2513,6 @@ namespace Ludots.Raylib.Render
             ApplySkyIbl(_frameLighting);
             ApplyFrameShadow(_shader, in _instancingShadowLocs);
         }
-
-        private void EnsureFrameLightingAppliedForSkinning()
-        {
-            if (_frameLighting == null)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} lit GpuSkinnedInstance requires {nameof(ApplyFrameLighting)} before draw.");
-            }
-
-            if (!_hasFrameViewPos)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} lit GpuSkinnedInstance requires camera view position before draw.");
-            }
-
-            EnsureSkinningShaderInitialized();
-            _frameLighting.Apply(_skinningShader, in _skinningLightingLocs);
-            _frameLighting.ApplyViewPosition(_skinningShader, in _skinningLightingLocs, _frameViewPos);
-            ApplyFrameShadow(_skinningShader, in _skinningShadowLocs);
-        }
-
         private void EnsureImmediateLitFrame()
         {
             if (_frameLighting == null)
@@ -3016,25 +2530,6 @@ namespace Ludots.Raylib.Render
             _immediateLit ??= new RaylibLitModel();
             _immediateLit.BeginFrame(_frameLighting, _frameViewPos, _frameShadow, _frameShadowTexelWorld);
         }
-
-        private static void RequireMeshNormals(in Mesh mesh, string lane)
-        {
-            if (mesh.normals == null)
-            {
-                throw new InvalidOperationException(
-                    $"{nameof(RaylibPrimitiveRenderer)} {lane} lit path requires mesh normals (vertexCount={mesh.vertexCount}); silent flat shading is forbidden.");
-            }
-        }
-
-        private static uint PackRgba(in Vector4 c)
-        {
-            uint r = Clamp01ToByte(c.X);
-            uint g = Clamp01ToByte(c.Y);
-            uint b = Clamp01ToByte(c.Z);
-            uint a = Clamp01ToByte(c.W);
-            return r | (g << 8) | (b << 16) | (a << 24);
-        }
-
         private static Color ToRaylibColor(in Vector4 c) => RaylibColorUtil.ToRaylibColor(in c);
 
         private Vector3 ResolveBillboardLitTintRgb()
@@ -3104,6 +2599,7 @@ namespace Ludots.Raylib.Render
 
             _decalRenderer.Dispose();
             _vegetationCutout.Dispose();
+            _gpuSkinned.Dispose();
 
             _gpuSkinnedModelCache.UnloadAll(model => _materialHostBinder?.DetachOwnedMaps(model));
             _gpuSkinnedModelCache.Dispose();
@@ -3128,11 +2624,6 @@ namespace Ludots.Raylib.Render
             RaylibShadowSampling.ClearTexture(ref _material);
             Rl.UnloadMaterial(_material);
             Rl.UnloadShader(_shader);
-            if (_skinningShaderReady)
-            {
-                Rl.UnloadShader(_skinningShader);
-                _skinningShaderReady = false;
-            }
             _initialized = false;
         }
 
@@ -3196,40 +2687,6 @@ namespace Ludots.Raylib.Render
                 Transforms = new RaylibMatrix[Math.Max(4, initialCapacity)];
                 Count = 0;
                 Revision = int.MinValue;
-            }
-
-            public void Add(in RaylibMatrix matrix)
-            {
-                if (Count >= Transforms.Length)
-                {
-                    Array.Resize(ref Transforms, Transforms.Length * 2);
-                }
-
-                Transforms[Count++] = matrix;
-            }
-        }
-
-        private readonly record struct GpuSkinnedInstanceBatchKey(
-            int MeshAssetId,
-            int MaterialId,
-            uint ColorKey,
-            int ClipIndex,
-            int FrameIndex);
-
-        private sealed class GpuSkinnedInstanceBatch
-        {
-            public readonly GpuSkinnedInstanceBatchKey Key;
-            public RaylibMatrix[] Transforms;
-            public int Count;
-            public Model Model;
-            public ModelAnimation* Animations;
-            public int AnimCount;
-            public bool BonesPrepared;
-
-            public GpuSkinnedInstanceBatch(GpuSkinnedInstanceBatchKey key, int initialCapacity = 256)
-            {
-                Key = key;
-                Transforms = new RaylibMatrix[Math.Max(4, initialCapacity)];
             }
 
             public void Add(in RaylibMatrix matrix)
